Package updates and reorganization of files
- Upgrade @11ty/eleventy to v3.0.0-alpha.10
- Upgrade @11ty/eleventy-img to v5.0.0-beta.1
- Move ./src/static to ./src/assets and update paths
- Organize the ./utils directory
