+++ .eslintrc.json
@@ -5,7 +5,7 @@
 		"es6": true
 	},
 	"parserOptions": {
-		"ecmaVersion": 2021
+		"ecmaVersion": 2022
 	},
 	"rules": {
         "indent": ["warn", 2],
@@ -16,7 +16,7 @@
     "ignorePatterns": ["!eleventy.config.js", "dist", "node_modules"],
     "overrides": [
         {
-            "files": ["src/static/js/*.js"],
+            "files": ["src/assets/scripts/*.js"],
             "env": {
                 "browser": true
             }
