+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "spencerharston.com",
-  "version": "2.0.0",
+  "version": "2.1.0",
   "description": "The personal website of Spencer Harston",
   "private": true,
   "scripts": {
@@ -8,10 +8,10 @@
     "lint:fix": "eslint \"./**/*.js\" --fix",
     "start": "run-p start:*",
     "start:eleventy": "eleventy --serve",
-    "start:postcss": "postcss src/styles/*.css --dir dist/styles --watch",
+    "start:postcss": "postcss src/assets/css/*.css --dir dist/assets/css --watch",
     "build": "run-s clean build:eleventy build:postcss",
     "build:eleventy": "eleventy",
-    "build:postcss": "postcss src/styles/*.css --dir dist/styles",
+    "build:postcss": "postcss src/assets/css/*.css --dir dist/assets/css",
     "clean": "rm -rf ./dist"
   },
   "repository": {
@@ -25,8 +25,8 @@
   },
   "homepage": "https://github.com/sphars/spencerharston.com#readme",
   "devDependencies": {
-    "@11ty/eleventy": "^2.0.1",
-    "@11ty/eleventy-img": "^3.1.8",
+    "@11ty/eleventy": "^3.0.0-alpha.10",
+    "@11ty/eleventy-img": "^5.0.0-beta.1",
     "@11ty/eleventy-plugin-rss": "^1.2.0",
     "@fontsource/arimo": "^5.0.19",
     "@fontsource/atkinson-hyperlegible": "^5.0.19",
