+++ src/feed.njk
@@ -23,7 +23,7 @@ metadata:
         <name>{{ metadata.author.name }}</name>
         <email>{{ metadata.author.email }}</email>
     </author>
-    <icon>/static/favicons/android-chrome-512x512.png</icon>
+    <icon>/assets/favicons/android-chrome-512x512.png</icon>
     {%- for post in collections.posts %}
     {% set absolutePostUrl %}{{ post.url | url | absoluteUrl(metadata.url) }}{% endset %}
     <entry>
