+++ src/index.njk
@@ -4,7 +4,7 @@ metaTagline: The personal website of Spencer Harston
 ---
 
 <div class="flex flex-col lg:flex-row items-center my-2">
-    <img class="rounded-full border border-gray-400 dark:border-gray-600 h-36 w-36 mb-2 lg:mb-0" src="/static/img/bird-crop.jpg" alt="photo of bird on a branch" title="pretty birdie">
+    <img class="rounded-full border border-gray-400 dark:border-gray-600 h-36 w-36 mb-2 lg:mb-0" src="/assets/img/bird-crop.jpg" alt="photo of bird on a branch" title="pretty birdie">
     <div class="flex flex-col items-center lg:items-start justify-center gap-3 py-0 px-4">
         <h1 class="text-3xl text-center font-bold">Spencer Harston</h1>
         <p class="text-center lg:text-left">My corner of the internet</p>
